Again fixing the controller configuration with many config files.
- fix pre-defined values for xinput and ds4
- changed datatypes in order to fix "no controller config"-found issues
with many config files

+++ utils/CPRetainVars.cpp
@@ -24,8 +24,8 @@ ControllerMapping gControllerMapping __attribute__((section(".data")));
 u8 gConfig_done __attribute__((section(".data"))) = 0;
 u8 gButtonRemappingConfigDone __attribute__((section(".data"))) = 0;
 
-u16 gHIDAttached __attribute__((section(".data"))) = 0;
-u16 gHIDCurrentDevice __attribute__((section(".data"))) = 0;
+u32 gHIDAttached __attribute__((section(".data"))) = 0;
+u32 gHIDCurrentDevice __attribute__((section(".data"))) = 0;
 
 u16 gHIDRegisteredDevices __attribute__((section(".data"))) = 0;
 
@@ -38,12 +38,12 @@ u8 gHID_Mouse_Mode __attribute__((section(".data"))) = HID_MOUSE_MODE_TOUCH;
 u32 gGamePadValues[CONTRPS_MAX_VALUE] __attribute__((section(".data")));
 
 u8  config_controller[gHIDMaxDevices][CONTRPS_MAX_VALUE][2] __attribute__((section(".data")));
-u16 config_controller_hidmask[gHIDMaxDevices] __attribute__((section(".data")));
+u32 config_controller_hidmask[gHIDMaxDevices] __attribute__((section(".data")));
 
-u16 gHID_LIST_GC __attribute__((section(".data"))) = 0;
-u16 gHID_LIST_DS3 __attribute__((section(".data"))) = 0;
-u16 gHID_LIST_KEYBOARD __attribute__((section(".data"))) = 0;
-u16 gHID_LIST_MOUSE __attribute__((section(".data"))) = 0;
+u32 gHID_LIST_GC __attribute__((section(".data"))) = 0;
+u32 gHID_LIST_DS3 __attribute__((section(".data"))) = 0;
+u32 gHID_LIST_KEYBOARD __attribute__((section(".data"))) = 0;
+u32 gHID_LIST_MOUSE __attribute__((section(".data"))) = 0;
 
 u16 gGamePadSlot __attribute__((section(".data"))) = 0;
 u16 gHID_SLOT_GC __attribute__((section(".data"))) = 0;
